services/auth_service/requirements.txt
Flask==3.0.0
pyjwt==2.9.0
